moveup_django_back/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.15,<3.16
django-cors-headers>=4.4,<4.6
django-filter>=24.3,<25
PyMySQL>=1.1,<1.2
